Move automation controls from operations to main page. This allows the play/pause emergency buttons be better reachable.
